edify: remove the need for fstab info during ota creation
change mount/format arguments to be only the partition and not depend on
the details of the filesystems. Those details can be determined in the
updater binary. This allows us to have an OTA that is compatible with both
NAND/ubifs and block/ext4.

We leave the TunePartition/WipeBlockDevice/WriteRawImage functions alone
as we don't use those and to keep those happy we allow it to inspect
the fstab_block file for details as we don't currently use these functions
and thus don't care.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>